Adds state for two rows of cell zoom views
