+++ src/components/spaces/cell-zoom-space.tsx
@@ -17,18 +17,20 @@ export class CellZoomSpaceComponent extends BaseComponent<IProps, IState> {
     const {cellZoom} = this.stores;
     const graphTitle = "Graph";
 
-    const cellGraphPanel1 = <Chart title="Chart Test" chartData={cellZoom.currentData} chartType={"horizontalBar"} />;
+    const cellGraphPanel1 = <Chart title="Chart Test" chartData={cellZoom.rows[0].currentData}
+      chartType={"horizontalBar"} />;
     const cellZoomComponent1 = <TwoUpDisplayComponent
       leftTitle="Investigate: Cell"
-      leftPanel={<OrganelleWrapper elementName="organelle-wrapper-1" />}
+      leftPanel={<OrganelleWrapper elementName="organelle-wrapper-1" rowIndex={0}/>}
       rightTitle={graphTitle}
       rightPanel={cellGraphPanel1}
     />;
 
-    const cellGraphPanel2 = <Chart title="Chart Test" chartData={cellZoom.currentData} chartType={"horizontalBar"} />;
+    const cellGraphPanel2 = <Chart title="Chart Test" chartData={cellZoom.rows[1].currentData}
+      chartType={"horizontalBar"} />;
     const cellZoomComponent2 = <TwoUpDisplayComponent
       leftTitle="Investigate: Cell"
-      leftPanel={<OrganelleWrapper elementName="organelle-wrapper-2" />}
+      leftPanel={<OrganelleWrapper elementName="organelle-wrapper-2" rowIndex={1}/>}
       rightTitle={graphTitle}
       rightPanel={cellGraphPanel2}
     />;
