+++ src/models/stores.ts
@@ -4,6 +4,7 @@ import { BackpackModel, BackpackModelType } from "./backpack";
 import { flatten } from "flat";
 import { CellZoomModel, CellZoomModelType } from "./spaces/cell-zoom/cell-zoom";
 import { CellMouseModel } from "../components/spaces/cell-zoom/cell-mouse";
+import { CellZoomRowModel } from "./spaces/cell-zoom/cell-zoom-row";
 
 export type Curriculum = "mouse";
 
@@ -30,6 +31,15 @@ export function createStores(params: ICreateStores, authoring: any): IStores {
       createPopulationsModel(authoring.curriculum, flatten(authoring.populations)),
     backpack: params && params.backpack || BackpackModel.create({}),
     cellZoom: params && params.cellZoom
-      || CellZoomModel.create({ organism: CellMouseModel.create({baseColor: "brown"}) })
+      || CellZoomModel.create({
+           organisms: {
+            one: CellMouseModel.create({ id: "one", baseColor: "brown" }),
+            two: CellMouseModel.create({ id: "two", baseColor: "white" })
+          },
+          rows: [
+            CellZoomRowModel.create({organism: "one"}),
+            CellZoomRowModel.create({organism: "two"}),
+          ]
+         })
   };
 }
